Repurpose old system settings page into a placeholder for new settings page
