+++ www/src/js/config.js
@@ -1,6 +1,5 @@
 module.exports = {
   PACKAGES: '/api/v2/packages/',
-  SETTINGS: '/api/v2/packages/ubuntu-core',
   FILTERED_SNAPS: [
     'snapweb',
     'ubuntu-core'
