--- www/src/js/models/system.js
@@ -1,10 +1,0 @@
-// system.js
-// system settings model
-
-var Backbone = require('backbone');
-var Marionette = require('backbone.marionette');
-var CONFIG = require('../config.js');
-
-module.exports = Backbone.Model.extend({
-  urlRoot: CONFIG.SETTINGS
-});
