+++ www/src/js/templates/settings.hbs
@@ -1,1 +1,5 @@
-<p>{{ name }} version: {{ version }}</p>
+<div class="inner-wrapper">
+  <div class="b-headline">
+    <h2>Settings</h2>
+  </div>
+</div>
