--- www/src/js/templates/system.hbs
@@ -1,3 +1,0 @@
-<h1>System settings</h1>
-<div class="region-system-settings"></div>
-<div class="region-installed"></div>
